src/test/java/magic/system/hyperion/components/RequirementMetaInformationTest.java
/*
 * The MIT License
 *
 * Copyright 2021 Thomas Lehmann.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 */
package magic.system.hyperion.components;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * Testing of class {@link RequirementMetaInformation}.
 *
 * @author Thomas Lehmann
 */
@DisplayName("Test of RequirementMetaInformation")
public class RequirementMetaInformationTest {
    /**
     * Test milestone.
     */
    private static final String TEST_MILESTONE = "1.0.0";

    /**
     * Test context.
     */
    private static final String TEST_CONTEXT = "task feature";

    @Test
    public void testDefault() {
        final var meta = new RequirementMetaInformation(TEST_MILESTONE, TEST_CONTEXT);
        assertEquals(TEST_MILESTONE, meta.getMilestone());
        assertEquals(TEST_CONTEXT, meta.getContext());
    }
}
